Make cffdrs singleton
