+++ api/app/data/ecodivision_seasons.py
@@ -5,35 +5,7 @@ import logging
 import geopandas
 from shapely.geometry import Point
 
-
-class Singleton:
-    """ Non thread-safe singleton decorator.
-        Stolen from: https://stackoverflow.com/a/7346105
-    """
-
-    def __init__(self, decorated):
-        self._decorated = decorated
-
-    def instance(self):
-        """
-        Returns the singleton instance. Upon its first call, it creates a
-        new instance of the decorated class and calls its `__init__` method.
-        On all subsequent calls, the already created instance is returned.
-
-        """
-        try:
-            return self._instance
-        except AttributeError:
-            # pylint: disable=attribute-defined-outside-init
-            self._instance = self._decorated()
-            return self._instance
-
-    def __call__(self):
-        raise TypeError('Singletons must be accessed through `instance()`.')
-
-    def __instancecheck__(self, inst):
-        return isinstance(inst, self._decorated)
-
+from app.utils.singleton import Singleton
 
 dirname = os.path.dirname(__file__)
 core_season_file_path = os.path.join(
